+++ packages/server/src/app/controllers/migration/migration.service.ts
@@ -4,13 +4,13 @@ import { ContentKind } from '@dragonfish/models/content';
 import { MigrationForm } from '@dragonfish/models/migration';
 
 import { OldBlogsService } from '../../db/blogs/blogs.service';
-import { ContentService } from '../../db/content';
+import { ContentStore } from '../../db/content';
 import { WorksService } from '../../db/works/works.service';
 
 @Injectable()
 export class MigrationService {
     constructor(
-        private readonly contentService: ContentService,
+        private readonly contentService: ContentStore,
         private readonly worksService: WorksService,
         private readonly blogsService: OldBlogsService,
     ) {}
