--- packages/server/src/app/controllers/migration/old-blogs.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('old-blogs')
-export class OldBlogsController {}
