--- packages/server/src/app/controllers/migration/old-works.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('old-works')
-export class OldWorksController {}
